fix: upgrade Flyway to 11.7.2 with PostgreSQL 18 support and fix schema mismatches
- Upgrade Flyway from 11.3.1 to 11.7.2 in parent pom
- Add flyway-database-postgresql dependency to all services
  (donation, followup, health, shelter, staff, supply)
- Migrate donation-service monetary fields from Double to BigDecimal
  (Donation, DonationHistory, DTOs, service layer, DataInitializer)
- Fix UserRepository.findByStatusIgnoreCase for enum type (use UserStatus)
- Add V3 migration for users.updated_at column
- Add PostgreSQL driver to health-service pom.xml
- Update README with all 10 services, corrected endpoints, and build instructions
- Add test-commands.txt with PostgreSQL and H2 profile curl commands

+++ donation-service/src/main/java/com/adoptapp/donationservice/dto/DonationHistoryResponse.java
@@ -1,5 +1,6 @@
 package com.adoptapp.donationservice.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record DonationHistoryResponse(
@@ -12,9 +13,9 @@ public record DonationHistoryResponse(
 
         String newStatus,
 
-        Double previousAmount,
+        BigDecimal previousAmount,
 
-        Double newAmount,
+        BigDecimal newAmount,
 
         String comment,
